major changes to extended_viz branch
